Automatically perform cleanup when running on testnet
Perform cleanup at the end of a testing session:

* withdraw rewards
* deregister stake addresses
* return funds to faucet
